Tidy up #1684
