+++ pkg/apis/application/v1alpha1/openapi_generated.go
@@ -105,6 +105,7 @@ func schema_pkg_apis_application_v1alpha1_AWSAuthConfig(ref common.ReferenceCall
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -358,6 +359,7 @@ func schema_pkg_apis_application_v1alpha1_ApplicationCondition(ref common.Refere
 				Required: []string{"type", "message"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -385,6 +387,7 @@ func schema_pkg_apis_application_v1alpha1_ApplicationDestination(ref common.Refe
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -691,7 +694,6 @@ func schema_pkg_apis_application_v1alpha1_ApplicationSourceKustomize(ref common.
 							Description: "CommonLabels adds additional kustomize commonLabels",
 							Type:        []string{"object"},
 							AdditionalProperties: &spec.SchemaOrBool{
-								Allows: true,
 								Schema: &spec.Schema{
 									SchemaProps: spec.SchemaProps{
 										Type:   []string{"string"},
@@ -725,6 +727,7 @@ func schema_pkg_apis_application_v1alpha1_ApplicationSourcePlugin(ref common.Ref
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -906,6 +909,7 @@ func schema_pkg_apis_application_v1alpha1_ApplicationSummary(ref common.Referenc
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1123,6 +1127,7 @@ func schema_pkg_apis_application_v1alpha1_Command(ref common.ReferenceCallback) 
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1181,6 +1186,7 @@ func schema_pkg_apis_application_v1alpha1_ComponentParameter(ref common.Referenc
 				Required: []string{"name", "value"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1270,6 +1276,7 @@ func schema_pkg_apis_application_v1alpha1_HealthStatus(ref common.ReferenceCallb
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1297,6 +1304,7 @@ func schema_pkg_apis_application_v1alpha1_HelmParameter(ref common.ReferenceCall
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1352,6 +1360,7 @@ func schema_pkg_apis_application_v1alpha1_HelmRepository(ref common.ReferenceCal
 				Required: []string{"url", "name"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1379,6 +1388,7 @@ func schema_pkg_apis_application_v1alpha1_InfoItem(ref common.ReferenceCallback)
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1405,6 +1415,7 @@ func schema_pkg_apis_application_v1alpha1_JWTToken(ref common.ReferenceCallback)
 				Required: []string{"iat"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1437,6 +1448,7 @@ func schema_pkg_apis_application_v1alpha1_JsonnetVar(ref common.ReferenceCallbac
 				Required: []string{"name", "value"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1469,6 +1481,7 @@ func schema_pkg_apis_application_v1alpha1_KsonnetParameter(ref common.ReferenceC
 				Required: []string{"name", "value"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1496,6 +1509,7 @@ func schema_pkg_apis_application_v1alpha1_KustomizeImageTag(ref common.Reference
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1782,6 +1796,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceActionDefinition(ref common.Re
 				Required: []string{"name", "action.lua"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1818,6 +1833,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceActionParam(ref common.Referen
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1906,6 +1922,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceDiff(ref common.ReferenceCallb
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1957,6 +1974,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceIgnoreDifferences(ref common.R
 				Required: []string{"group", "kind", "jsonPointers"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -1971,7 +1989,6 @@ func schema_pkg_apis_application_v1alpha1_ResourceNetworkingInfo(ref common.Refe
 						SchemaProps: spec.SchemaProps{
 							Type: []string{"object"},
 							AdditionalProperties: &spec.SchemaOrBool{
-								Allows: true,
 								Schema: &spec.Schema{
 									SchemaProps: spec.SchemaProps{
 										Type:   []string{"string"},
@@ -1997,7 +2014,6 @@ func schema_pkg_apis_application_v1alpha1_ResourceNetworkingInfo(ref common.Refe
 						SchemaProps: spec.SchemaProps{
 							Type: []string{"object"},
 							AdditionalProperties: &spec.SchemaOrBool{
-								Allows: true,
 								Schema: &spec.Schema{
 									SchemaProps: spec.SchemaProps{
 										Type:   []string{"string"},
@@ -2173,6 +2189,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceOverride(ref common.ReferenceC
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2222,6 +2239,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceRef(ref common.ReferenceCallba
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2290,6 +2308,7 @@ func schema_pkg_apis_application_v1alpha1_ResourceResult(ref common.ReferenceCal
 				Required: []string{"group", "version", "kind", "namespace", "name"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2483,6 +2502,7 @@ func schema_pkg_apis_application_v1alpha1_SyncOperationResource(ref common.Refer
 				Required: []string{"kind", "name"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2566,6 +2586,7 @@ func schema_pkg_apis_application_v1alpha1_SyncPolicyAutomated(ref common.Referen
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2646,6 +2667,7 @@ func schema_pkg_apis_application_v1alpha1_SyncStrategyApply(ref common.Reference
 				},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2716,6 +2738,7 @@ func schema_pkg_apis_application_v1alpha1_TLSClientConfig(ref common.ReferenceCa
 				Required: []string{"insecure"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
 
@@ -2736,5 +2759,6 @@ func schema_pkg_apis_application_v1alpha1_objectMeta(ref common.ReferenceCallbac
 				Required: []string{"Name"},
 			},
 		},
+		Dependencies: []string{},
 	}
 }
