+++ test/e2e/app_management_test.go
@@ -8,13 +8,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/argoproj/argo-cd/errors"
-
 	"github.com/stretchr/testify/assert"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/argoproj/argo-cd/common"
+	"github.com/argoproj/argo-cd/errors"
 	. "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
 	argorepo "github.com/argoproj/argo-cd/reposerver/repository"
 	"github.com/argoproj/argo-cd/server/application"
@@ -371,49 +370,6 @@ func TestKsonnetApp(t *testing.T) {
 			}
 			assert.Equal(t, serviceType, "LoadBalancer")
 		})
-}
-
-func TestKustomize2AppSource(t *testing.T) {
-
-	patchLabelMatchesFor := func(kind string) func(app *Application) {
-		return func(app *Application) {
-			name := "k2-patched-guestbook-ui"
-			labelValue, err := fixture.Run(
-				"", "kubectl", "-n="+fixture.DeploymentNamespace(),
-				"get", kind, name,
-				"-ojsonpath={.metadata.labels.patched-by}")
-			assert.NoError(t, err)
-			assert.Equal(t, "argo-cd", labelValue, "wrong value of 'patched-by' label of %s %s", kind, name)
-		}
-	}
-
-	Given(t).
-		Path(guestbookPath).
-		NamePrefix("k2-").
-		When().
-		Create().
-		Refresh(RefreshTypeHard).
-		PatchApp(`[
-			{
-				"op": "replace",
-				"path": "/spec/source/kustomize/namePrefix",
-				"value": "k2-patched-"
-			},
-			{
-				"op": "add",
-				"path": "/spec/source/kustomize/commonLabels",
-				"value": {
-					"patched-by": "argo-cd"
-				}
-			}
-		]`).
-		Then().
-		Expect(Success("")).
-		When().
-		Sync().
-		Then().
-		And(patchLabelMatchesFor("Service")).
-		And(patchLabelMatchesFor("Deployment"))
 }
 
 const actionsConfig = `discovery.lua: return { sample = {} }
